docker/room_api/requirements.txt
flask
line-bot-sdk
uvicorn
python-dotenv
pymongo